pyproject.toml
[tool.poetry]
name = "python-genai"
version = "0.1.0"
description = ""
authors = ["Muhammad Zafeer <zafeerwains@gmail.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.3.21"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.8"
langchain-openai = "^0.2.10"
langchain = "^0.3.10"
langchain-google-genai = "^2.0.6"
fastapi = {extras = ["standard"], version = "^0.115.11"}
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pypdf = "^5.1.0"
cryptography = "^44.0.0"
langgraph = "^0.2.59"
ipython = "^9.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
